chore(typo): rename hook files
