+++ frontend/src/modules/hooks/useIsMobileDevice .ts
